# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/": {
            "get": {
                "description": "Returns all stored metrics in an HTML table.",
                "consumes": [
                    "text/html"
                ],
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "Metrics"
                ],
                "summary": "Get All Metrics",
                "operationId": "get-all-metrics",
                "responses": {
                    "200": {
                        "description": "\u003c!DOCTYPE html\u003e...",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/ping": {
            "get": {
                "description": "Pings the database to check its availability.",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Health"
                ],
                "summary": "Health Check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/update": {
            "post": {
                "description": "Updates or adds metrics received in request and returns the updated metrics.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Metrics"
                ],
                "summary": "Update or add metrics from JSON",
                "operationId": "update-metric-from-json",
                "parameters": [
                    {
                        "description": "Metrics data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/metrics.Metric"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated metric",
                        "schema": {
                            "$ref": "#/definitions/metrics.Metric"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/update/{metricType}/{metricName}/{metricValue}": {
            "post": {
                "description": "Updates the specified metric with the provided value.",
                "consumes": [
                    "text/plain"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Metrics"
                ],
                "summary": "Update Metric",
                "operationId": "update-metric",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Name of the metric to update",
                        "name": "metricName",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Type of the metric to update",
                        "name": "metricType",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Value to update the metric with",
                        "name": "metricValue",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated metric value",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/updates": {
            "post": {
                "description": "Updates or adds metrics received in request and returns the updated metrics.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Metrics"
                ],
                "summary": "Update or add metrics from JSON",
                "operationId": "update-metrics-from-json",
                "parameters": [
                    {
                        "description": "Metrics data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/metrics.Metric"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated metrics",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/metrics.Metric"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/value": {
            "post": {
                "description": "Gets requested metric by id and type and returns it's id, type and value in JSON format.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Metrics"
                ],
                "summary": "Get Metric JSON",
                "operationId": "get-metric-json",
                "parameters": [
                    {
                        "description": "Metric data",
                        "name": "metric",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/metrics.Metric"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/value/{metricType}/{metricName}": {
            "get": {
                "description": "searches metric by id and type and returns it's value in plain text.",
                "consumes": [
                    "text/plain"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Metrics"
                ],
                "summary": "Get Metric",
                "operationId": "get-metric",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Name of the metric to retrieve",
                        "name": "metricName",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Type of the metric to retrieve",
                        "name": "metricType",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mertic value",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "metrics.Metric": {
            "type": "object",
            "properties": {
                "delta": {
                    "description": "Delta metric value used for Counter metric type",
                    "type": "integer"
                },
                "id": {
                    "description": "ID metric name",
                    "type": "string"
                },
                "type": {
                    "description": "MType metric type (Gauge or Counter)",
                    "type": "string"
                },
                "value": {
                    "description": "Value metric value used for Gauge metric type",
                    "type": "number"
                }
            }
        }
    }
}